Introduce /widget route and refactor VerificationForm to a new component
