+++ app/root.tsx
@@ -1,4 +1,4 @@
-import { isRouteErrorResponse, Links, Meta, Outlet, Scripts, ScrollRestoration, Link } from "react-router";
+import { isRouteErrorResponse, Links, Meta, Outlet, Scripts, ScrollRestoration, Link, useLocation } from "react-router";
 import { FaGithub } from "react-icons/fa";
 
 import type { Route } from "./+types/root";
@@ -10,6 +10,9 @@ import { CompilerVersionsProvider } from "./contexts/CompilerVersionsContext";
 export const links: Route.LinksFunction = () => [];
 
 export function Layout({ children }: { children: React.ReactNode }) {
+  const location = useLocation();
+  const isWidgetRoute = location.pathname === '/widget';
+
   return (
     <html lang="en">
       <head>
@@ -29,29 +32,31 @@ export function Layout({ children }: { children: React.ReactNode }) {
         <ServerConfigProvider>
           <ChainsProvider>
             <CompilerVersionsProvider>
-              <header className="shadow-sm">
-                <div className="mx-auto py-3 md:py-4 flex items-center justify-between w-full max-w-[100rem] px-4 md:px-12 lg:px-12 xl:px-24">
-                  <Link to="/" className="flex items-center">
-                    <img
-                      src="/sourcify.png"
-                      alt="Sourcify Logo"
-                      className="h-8 md:h-10 w-auto mr-2 md:mr-3"
-                      width={32}
-                      height={32}
-                    />
-                    <span className="text-gray-700 font-vt323 text-xl md:text-2xl">verify.sourcify.eth</span>
-                  </Link>
-                  <a
-                    href="https://github.com/sourcifyeth/verify.sourcify.dev"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    className="text-gray-600 hover:text-gray-900 transition-colors p-2"
-                    aria-label="View source code on GitHub"
-                  >
-                    <FaGithub className="w-6 h-6" />
-                  </a>
-                </div>
-              </header>
+              {!isWidgetRoute && (
+                <header className="shadow-sm">
+                  <div className="mx-auto py-3 md:py-4 flex items-center justify-between w-full max-w-[100rem] px-4 md:px-12 lg:px-12 xl:px-24">
+                    <Link to="/" className="flex items-center">
+                      <img
+                        src="/sourcify.png"
+                        alt="Sourcify Logo"
+                        className="h-8 md:h-10 w-auto mr-2 md:mr-3"
+                        width={32}
+                        height={32}
+                      />
+                      <span className="text-gray-700 font-vt323 text-xl md:text-2xl">verify.sourcify.eth</span>
+                    </Link>
+                    <a
+                      href="https://github.com/sourcifyeth/verify.sourcify.dev"
+                      target="_blank"
+                      rel="noopener noreferrer"
+                      className="text-gray-600 hover:text-gray-900 transition-colors p-2"
+                      aria-label="View source code on GitHub"
+                    >
+                      <FaGithub className="w-6 h-6" />
+                    </a>
+                  </div>
+                </header>
+              )}
               <main>{children}</main>
             </CompilerVersionsProvider>
           </ChainsProvider>
